+++ tests/5_set.c
@@ -1,5 +1,29 @@
-#include <cstd/CLog.h>
+/*
+ * MIT License
+ *
+ * Copyright (c) 2024 Subhadip Roy Chowdhury
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 #include <cstd/CHashSet.h>
+#include <cstd/CLog.h>
 
 #include <assert.h>
 #include <stdint.h>
@@ -57,8 +81,8 @@ int main() {
     enable_location();
     shortened_location();
     CLog(INFO, "create_hash_set()");
-    CHashSet_t* set;
-    CResult_t *res = CHashSet_new(20,int_compare, int_hash, free);
+    CHashSet_t *set;
+    CResult_t *res = CHashSet_new(20, int_compare, int_hash, free);
     assert(!CResult_is_error(res));
     set = CResult_get(res);
     test_insert(set);